Move `make_pip_conf`, `pip_conf` and `pip_with_index_conf` fixtures to `tests/conftest.py`.
