+++ tests/test_cli_compile.py
@@ -16,59 +16,6 @@ from piptools.repositories import PyPIRepository
 from piptools.scripts.compile import cli
 
 
-@pytest.fixture
-def make_pip_conf(tmpdir, monkeypatch):
-    created_paths = []
-
-    def _make_pip_conf(content):
-        pip_conf_file = "pip.conf" if os.name != "nt" else "pip.ini"
-        path = (tmpdir / pip_conf_file).strpath
-
-        with open(path, "w") as f:
-            f.write(content)
-
-        monkeypatch.setenv("PIP_CONFIG_FILE", path)
-
-        created_paths.append(path)
-        return path
-
-    try:
-        yield _make_pip_conf
-    finally:
-        for path in created_paths:
-            os.remove(path)
-
-
-@pytest.fixture
-def pip_conf(make_pip_conf):
-    return make_pip_conf(
-        dedent(
-            """\
-            [global]
-            no-index = true
-            find-links = {wheels_path}
-            """.format(
-                wheels_path=MINIMAL_WHEELS_PATH
-            )
-        )
-    )
-
-
-@pytest.fixture
-def pip_with_index_conf(make_pip_conf):
-    return make_pip_conf(
-        dedent(
-            """\
-            [global]
-            index-url = http://example.com
-            find-links = {wheels_path}
-            """.format(
-                wheels_path=MINIMAL_WHEELS_PATH
-            )
-        )
-    )
-
-
 def test_default_pip_conf_read(pip_with_index_conf, runner):
     # preconditions
     with open("requirements.in", "w"):
